refactor to move workflow classes to own file
